pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "yelp-rag-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "gradio>=4.0.0",
    "langchain",
    "langchain-core",
    "langchain-ollama",
    "langgraph",
    "sentence-transformers",
    "faiss-cpu",
    "torch",
    "transformers",
    "numpy",
    "pandas",
    "requests",
    "pyyaml",
]

[tool.setuptools.packages.find]
where = ["src"]
